# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ashva Experts API (Consolidated)
  version: 0.2.0
  description: |
    Consolidated API specification for manual testing during development.

    Notes:
    - Base URL for local testing via Nginx: `http://localhost:8080`
    - All `/api/v1/*` routes are served through the gateway.
    - Auth endpoints under `/api/v1/auth/*` are public.
    - Non-auth endpoints require `Authorization: Bearer <access_token>`.
servers:
  - url: http://localhost:8080
    description: Local via Nginx (recommended)
  - url: http://localhost:8000
    description: Local gateway direct
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/content/case-studies:
    get:
      tags: [content]
      summary: List published case studies (public)
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: OK
    post:
      tags: [content]
      summary: Create case study (protected)
      description: Content management is restricted to `admin` and `cms_user`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            example:
              title: "Success Story: Acme Corp"
              slug: success-story-acme-corp
              summary: Brief summary
              content: Full content here...
              featured_image_url: https://example.com/image.jpg
              seo_title: "Acme Corp Success Story"
              seo_description: Learn how we helped...
              seo_keywords: case study, success, acme
              og_image_url: https://example.com/og.jpg
      responses:
        "201":
          description: Created
  /api/v1/content/case-studies/{slug}:
    get:
      tags: [content]
      summary: Get case study by slug (public)
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
  /api/v1/content/manage/case-studies:
    get:
      tags: [content]
      summary: List case studies (manage) (protected)
      description: Content management is restricted to `admin` and `cms_user`.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [draft, published, archived] }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: OK
  /api/v1/content/case-studies/{case_study_id}:
    patch:
      tags: [content]
      summary: Update case study (protected)
      description: Slug is immutable after publish.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: case_study_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            example:
              title: Updated title
              summary: Updated summary
      responses:
        "200":
          description: OK
    delete:
      tags: [content]
      summary: Delete case study (protected)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: case_study_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
  /api/v1/content/case-studies/{case_study_id}/publish:
    post:
      tags: [content]
      summary: Publish case study (protected)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: case_study_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
  /api/v1/content/case-studies/{case_study_id}/unpublish:
    post:
      tags: [content]
      summary: Unpublish case study back to draft (protected)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: case_study_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
...